create authority_per_epoch_store_utils and put cost related structs/functions there for better orgnization
